requirements.txt
tesseract==0.1.3
pytesseract==0.3.10
pandas==1.5.2
nltk==3.8.1
spacy==2.3.9
pyresparser==1.0.6
datetime==5.0
streamlit==1.17.0
openai==0.26.5
streamlit_chat==0.0.2.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
python-docx==0.8.11
emoji==2.2.0
